// src/model/jsonApiBodyLogin.ts
/**
 * Microservicio orquestador
 * microservicio orquestador para conectar los microservicios de persona, oferta y negocio
 *
 * OpenAPI spec version: 1.0.0
 * Contact: alexis.correa@accenture.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { LoginRequest } from './loginRequest';


export class JsonApiBodyLogin {
    persona: Array<LoginRequest>;
    constructor(){
        
    }
}